requirements.txt
# Alice AI - Requirements
# Python 3.11.0

google-genai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
GitPython>=3.1.40
loguru>=0.7.2
typing_extensions>=4.8.0
